// src/app/modules/configuration/class-schedules/model/holiday-search-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface HolidaySearchView {
    classId?: Array<number>;
    sectionId?: Array<number>;
    fromDate?: string;
    fromDateBegin?: string;
    fromDateEnd?: string;
    toDate?: string;
    toDateBegin?: string;
    toDateEnd?: string;
    holidayReason?: Array<any>;
    isStaffHoliday?: Array<any>;
    sortOrder: 0 | 1 | -1;
    sortBy: any;
    pageNumber: number;
    pageSize: number;
}